Sync with local: Enable prop insertion/deletion.
The user can now click on a ground tile together with a modifier key to add a new prop onto that tile; and similarly to remove the prop by modifier-clicking on it. Can't yet change the type of a prop, though, and all props added are (and thus remain) trees, by default.