Fix placing unreachable props at track perimeter.
Enforce a margin of no-placement for props around the track's perimeter, to prevent the props from becoming inaccessible in the editor's 3d view. This is kind of a hack, since ideally you should be able to move the camera past the track's boundaries and interact with props that're outside the perimeter (the game can still display out-of-bounds props just off the near-end z axis, so you should be able to place them there); but at the moment, the camera is bound to the track's extent and thus props can become obscured if not herded.