Shared mode: Disable Fetch's HTTP cache
So  you get a fresh copy of the server-side project files on page load.
This is important, since edits made by other participants that are sent
from the server to the client only update the edited portions of the
project data - so the client needs to have a full fresh copy of the
project data when they start their editing session.