Fully rewrite loading of project data
The server now always provides projects and related metadata as JSON
responses to GET requests from the client; rather than being a mixture of
zip files, raw binary, JSON, etc. that the client fetches directly as files.
This was done in preparation for a proper separation of the app into the
client, the server, and the server-side database.

This change necessitated large changes in several facets of the code-base
that had to do with project resources. Some of these changes are still
work-in-progress; and certain features of the program are currently in a
broken state, to be fixed in the commits following.