Renderer: Update to the latest version
Adds the latest version of the retro n-gon renderer.

The most notable change with the renderer is that texture pixel data is
now accessed inline rather than via a function call, to which
RallySportED's re-implementation of the renderer's texture() object needs
to adapt.

Another change with this version of the renderer is that it performs a bit
better than the previous version. It's still not as fast as RallySportED's
first-gen render engine, but it's getting there.