Implement a scene system to replace the current view system
Views/scenes are views to the current project - e.g. a 3d scene which
shows a textured 3d mesh of the project's track.

The new scene system lets each scene define its own render and input-
handling routines, which hopefully makes it easier to extend the program
with new kinds of scenes.

Note that this commit temporarily disables the top-down 3d view. It's not
yet known how long it'll remain disabled - but I maybe want to get rid of
it completely as a separate scene, and instead implement something like
mouse wheel scrolling to move/tilt the camera up and down for a gradient
between the normal 3d view and the top-down view.