Import latest version of the retro n-gon renderer
This fixes a bug in rendering where, when the wireframe is disabled, black
horizontal lines remain on certain polygons.

Note that to fix this bug in the retro n-gon renderer, modifications were
made to it with RallySportED-js in mind - and not necessarily the correct
functioning of the renderer in general. It remains unclear whether the
bug was inherent in the renderer or in RallySportED-js; and the bug is
thus still considered to be open.