Reset mouse buttons' status when cursor leaves canvas
To prevent the buttons from sticking.

This isn't a perfect solution in the sense that if the cursor does leave
the canvas while the user is holding down a button, the user will then
need to re-press the button when the cursor is back over the canvas in
order for the press/hold to be registered again.

In relation to the above, one could simply have the canvas grab mouse
focus (via the Pointer Lock API) and thus not have to worry about the
cursor leaving its designated area. But I'd rather avoid this if possible,
since I think it feels better for usability to not have the cursor locked
to the app during use.