Add vertical camera zoom with Shift + wheel
This commit re-introduces the top-down view that was removed in an earlier
commit. But instead of it being a discrete view mode, the user can now use
the mouse wheel to move the camera along a gradient between the normal,
straight-on view and a top-down view.

N.B.: The wheel event delta varies from browser to browser - the current
code is tuned for Chrome 7x and is known to produce suboptimal results in
at least some versions of Firefox (including new ones). Future commits
will likely introduce a more stable way of handling the delta cross-
browser.