Fix shared-mode editing
Shared-mode editing had broken during the current re-write of the
codebase. This commit re-introduces workable shared-mode editing.

Note that this fix has not yet been extensively tested - it's possible
that some issues remain. One is encouraged to not yet consider this
implementation production-ready (although the app as a whole is only
in beta, anyway).