Add a basic test for whether a mouse and keyboard are available
A mouse and keyboard are currently required by the RallySportED-js UI,
so it's useful to test for their presence and warn the user if not
available.

Currently, this just checks the user agent string to see if the user
might be on mobile and therefore without mouse/keyboard, but of course
this isn't a foolproof test - it's just a first approximation.