manifesto export: Use command #3 to create and define each prop
Rather than a combination of commands #4/#5. The idea now is that the
manifesto file must explicitly define each of the track's props, rather
than only reporting modifications to the existing props.

This change also bumps the required loader version to 5 - tracks saved
with this editor are no longer compatible with earlier versions of the
loader.