Move app's runtime files into the rallysported-js/ directory
All files needed to run RallySportED-js are now in the rallysported-js/
directory. Auxiliary files, which may operate on the app (e.g. unit
tests) but which the app doesn't depend on, remain in the repo's root.

This is to make it easier to copy the app onto a server, and in general
to more clearly delineate which files the app needs and which are extras.