Remove shared editing
This implementation of shared editing wasn't well integrated into the
rest of the codebase.

Since the project is about to get an overhaul to the way it handles
track data locality, it's best to get the shared editing guff out of
the way at least until the transition is done.

Whether shared editing will be re-added is to be seen. It's a cool
feature but wasn't super used.