Implement UI components
UI components are reusable visual elements - like a minimap, a string
label, etc. - that can be added to a scene's UI. Each component defines
a draw() function that renders the component into the currently-active
canvas, and an update() function that mutates the component and the
scene's state in response to user input.