scenes/3d: Use Rally-Sport's one-point perspective projection
Instead of proper 3D perspective.

Since the game uses texture-mapping that's dependent on viewing angle,
it's important that the editor has the exact same angle or texturing
will look different (sometimes very different).

Note that we're not yet doing proper frustum clipping in this one-
point projection, which means objects at the edges of the screen will
be misshapen.