Implement playing the current track in a browser DOSBox
The UI's menu bar now includes two new buttons: Test and Play. They both
launch a browser version of DOSBox (js-box) that loads up Rally-Sport
and the track's data, so the user can play the track in-game.

The commit also introduces the URL hash "#play", which causes the
given track to be loaded into the browser DOSBox immediately on
page load.