Scale rendering by screen width instead of constant value
Instead of using the 0.25 constant, the render scale value is now
adjusted at run-time to be a ratio of the current screen (browser
window) width.

This means that screen size (not accounting for aspect ratio) won't
(or shouldn't) affect the size of the render viewport. E.g. the same
number of ground tiles - and of the same relative size - will be
visible in the terrain editor view regardless of the size of the
screen; larger screens will just have larger pixels (smaller render
scale) to fill the extra screen space.

I don't have screens above 1080p to test on, but let's just assume
for now that it works as intended.